+++ twilight-model/src/application/interaction/application_command/option.rs
@@ -83,16 +83,14 @@ impl<'de> Deserialize<'de> for CommandDataOption {
             Focused,
         }
 
-        // Id before string such that IDs will always be interpreted
-        // as such, this does mean that string inputs that looks like
-        // IDs will have to be caught if it is a string.
+        // An `Id` variant is purposely not present here to prevent wrongly
+        // parsing string options as numbers, trimming leading zeroes.
         #[derive(Debug, Deserialize)]
         #[serde(untagged)]
         enum ValueEnvelope {
             Boolean(bool),
             Integer(i64),
             Number(f64),
-            Id(Id<GenericMarker>),
             String(String),
         }
 
@@ -102,7 +100,6 @@ impl<'de> Deserialize<'de> for CommandDataOption {
                     Self::Boolean(b) => Unexpected::Bool(*b),
                     Self::Integer(i) => Unexpected::Signed(*i),
                     Self::Number(f) => Unexpected::Float(*f),
-                    Self::Id(_) => Unexpected::Other("ID"),
                     Self::String(s) => Unexpected::Str(s),
                 }
             }
@@ -114,7 +111,6 @@ impl<'de> Deserialize<'de> for CommandDataOption {
                     Self::Boolean(b) => Display::fmt(b, f),
                     Self::Integer(i) => Display::fmt(i, f),
                     Self::Number(n) => Display::fmt(n, f),
-                    Self::Id(i) => Display::fmt(i, f),
                     Self::String(s) => Display::fmt(s, f),
                 }
             }
@@ -202,8 +198,10 @@ impl<'de> Deserialize<'de> for CommandDataOption {
                         CommandOptionType::Attachment => {
                             let val = value_opt.ok_or_else(|| DeError::missing_field("value"))?;
 
-                            if let ValueEnvelope::Id(id) = val {
-                                CommandOptionValue::Attachment(id.cast())
+                            if let ValueEnvelope::String(id) = &val {
+                                CommandOptionValue::Attachment(id.parse().map_err(|_| {
+                                    DeError::invalid_type(val.as_unexpected(), &"attachment id")
+                                })?)
                             } else {
                                 return Err(DeError::invalid_type(
                                     val.as_unexpected(),
@@ -223,8 +221,10 @@ impl<'de> Deserialize<'de> for CommandDataOption {
                         CommandOptionType::Channel => {
                             let val = value_opt.ok_or_else(|| DeError::missing_field("value"))?;
 
-                            if let ValueEnvelope::Id(id) = val {
-                                CommandOptionValue::Channel(id.cast())
+                            if let ValueEnvelope::String(id) = &val {
+                                CommandOptionValue::Channel(id.parse().map_err(|_| {
+                                    DeError::invalid_type(val.as_unexpected(), &"channel id")
+                                })?)
                             } else {
                                 return Err(DeError::invalid_type(
                                     val.as_unexpected(),
@@ -244,8 +244,10 @@ impl<'de> Deserialize<'de> for CommandDataOption {
                         CommandOptionType::Mentionable => {
                             let val = value_opt.ok_or_else(|| DeError::missing_field("value"))?;
 
-                            if let ValueEnvelope::Id(id) = val {
-                                CommandOptionValue::Mentionable(id)
+                            if let ValueEnvelope::String(id) = &val {
+                                CommandOptionValue::Mentionable(id.parse().map_err(|_| {
+                                    DeError::invalid_type(val.as_unexpected(), &"mentionable id")
+                                })?)
                             } else {
                                 return Err(DeError::invalid_type(
                                     val.as_unexpected(),
@@ -278,8 +280,10 @@ impl<'de> Deserialize<'de> for CommandDataOption {
                         CommandOptionType::Role => {
                             let val = value_opt.ok_or_else(|| DeError::missing_field("value"))?;
 
-                            if let ValueEnvelope::Id(id) = val {
-                                CommandOptionValue::Role(id.cast())
+                            if let ValueEnvelope::String(id) = &val {
+                                CommandOptionValue::Role(id.parse().map_err(|_| {
+                                    DeError::invalid_type(val.as_unexpected(), &"role id")
+                                })?)
                             } else {
                                 return Err(DeError::invalid_type(val.as_unexpected(), &"role id"));
                             }
@@ -287,17 +291,10 @@ impl<'de> Deserialize<'de> for CommandDataOption {
                         CommandOptionType::String => {
                             let val = value_opt.ok_or_else(|| DeError::missing_field("value"))?;
 
-                            match val {
-                                ValueEnvelope::String(s) => CommandOptionValue::String(s),
-                                ValueEnvelope::Id(id) => {
-                                    CommandOptionValue::String(id.get().to_string())
-                                }
-                                other => {
-                                    return Err(DeError::invalid_type(
-                                        other.as_unexpected(),
-                                        &"string",
-                                    ));
-                                }
+                            if let ValueEnvelope::String(s) = val {
+                                CommandOptionValue::String(s)
+                            } else {
+                                return Err(DeError::invalid_type(val.as_unexpected(), &"string"));
                             }
                         }
                         CommandOptionType::SubCommand => CommandOptionValue::SubCommand(options),
@@ -307,8 +304,10 @@ impl<'de> Deserialize<'de> for CommandDataOption {
                         CommandOptionType::User => {
                             let val = value_opt.ok_or_else(|| DeError::missing_field("value"))?;
 
-                            if let ValueEnvelope::Id(id) = val {
-                                CommandOptionValue::User(id.cast())
+                            if let ValueEnvelope::String(id) = &val {
+                                CommandOptionValue::User(id.parse().map_err(|_| {
+                                    DeError::invalid_type(val.as_unexpected(), &"user id")
+                                })?)
                             } else {
                                 return Err(DeError::invalid_type(val.as_unexpected(), &"user id"));
                             }
@@ -681,6 +680,31 @@ mod tests {
                 Token::U8(CommandOptionType::Number as u8),
                 Token::Str("value"),
                 Token::String("1"),
+                Token::StructEnd,
+            ],
+        );
+    }
+
+    #[test]
+    fn leading_zeroes_string_option_value() {
+        let value = CommandDataOption {
+            name: "opt".to_string(),
+            value: CommandOptionValue::String("0001".to_owned()),
+        };
+
+        serde_test::assert_de_tokens(
+            &value,
+            &[
+                Token::Struct {
+                    name: "CommandDataOption",
+                    len: 3,
+                },
+                Token::Str("name"),
+                Token::Str("opt"),
+                Token::Str("type"),
+                Token::U8(CommandOptionType::String as u8),
+                Token::Str("value"),
+                Token::String("0001"),
                 Token::StructEnd,
             ],
         );